+++ src/Component.c
@@ -1,7 +1,7 @@
 #include "Component.h"
 #include <string.h>
 
-int ComponentInit(Component* component, uint32_t objectId) {
+int ComponentInit(Component* component, uint64_t objectId) {
 	memset(component, 0, sizeof(Component));
 	component->objId = objectId;
 	return 0;
