+++ src/Components/ComponentEventListener.c
@@ -2,7 +2,7 @@
 #include "../Error.h"
 #include <string.h>
 
-int EventListenerComponentInit(EventListenerComponent* evListener, uint32_t objectId) {
+int EventListenerComponentInit(EventListenerComponent* evListener, uint64_t objectId) {
 	memset(evListener, 0, sizeof(EventListenerComponent));
 	return ComponentInit((Component*)evListener, objectId);
 }
