+++ src/Components/ComponentPhysics.c
@@ -5,7 +5,7 @@
 #include <stdio.h>
 #include <string.h>
 
-int PhysicsComponentInit(PhysicsComponent* phy, uint32_t objectId) {
+int PhysicsComponentInit(PhysicsComponent* phy, uint64_t objectId) {
 	memset(phy, 0, sizeof(PhysicsComponent));
 	PROPAGATE_ERROR(ComponentInit((Component*)phy, objectId));
 	return 0;
@@ -21,8 +21,8 @@ void PhysicsComponentDeinit(PhysicsComponent* phy) {
 
 void PhysicsComponentContactCB(Box2DContact* contact) {
 	Level* level = CurrentLevel();
-	uint32_t phyIdA = (uint32_t) ((uintptr_t) Box2DBodyGetUserData(Box2DFixtureGetBody(Box2DContactGetFixtureA(contact))));
-	uint32_t phyIdB = (uint32_t) ((uintptr_t) Box2DBodyGetUserData(Box2DFixtureGetBody(Box2DContactGetFixtureB(contact))));
+	uint64_t phyIdA = (uint64_t) ((uintptr_t) Box2DBodyGetUserData(Box2DFixtureGetBody(Box2DContactGetFixtureA(contact))));
+	uint64_t phyIdB = (uint64_t) ((uintptr_t) Box2DBodyGetUserData(Box2DFixtureGetBody(Box2DContactGetFixtureB(contact))));
 	PhysicsComponent* phyA = BucketGetById(&level->physics, phyIdA);
 	PhysicsComponent* phyB = BucketGetById(&level->physics, phyIdB);
 	if (phyA && phyB) {
