+++ src/InsertionList.h
@@ -6,17 +6,17 @@
 
 typedef struct _InsertionList {
 	Array array;
-	int (*comparator)(uint32_t, uint32_t);
+	int (*comparator)(uint64_t, uint64_t);
 } InsertionList;
 
-int InsertionListInit(InsertionList* list, size_t maxItemCount, int (*comparator)(uint32_t, uint32_t));
+int InsertionListInit(InsertionList* list, size_t maxItemCount, int (*comparator)(uint64_t, uint64_t));
 void InsertionListDeinit(InsertionList* list);
 
 size_t InsertionListLength(InsertionList* list);
-uint32_t InsertionListGet(InsertionList* list, size_t i);
+uint64_t InsertionListGet(InsertionList* list, size_t i);
 
-void InsertionListInsert(InsertionList* list, uint32_t id);
-void InsertionListRemove(InsertionList* list, uint32_t id);
+void InsertionListInsert(InsertionList* list, uint64_t id);
+void InsertionListRemove(InsertionList* list, uint64_t id);
 void InsertionListSort(InsertionList* list);
 
 #endif
