+++ src/Objects/ObjectTile.c
@@ -9,7 +9,7 @@ int ObjectTileInit(Object* obj, TileDef tileDef, Vec2F position) {
 	PROPAGATE_ERROR(ObjectInit(obj, position));
 
 	if (tileDef.colliderSize.x && tileDef.colliderSize.y) {
-		uint32_t phyId = 0;
+		uint64_t phyId = 0;
 		PhysicsComponent* phy = ObjectAddAndInitPhysics(obj, &phyId);
 		phy->body = Box2DUtilsCreateStaticBox(phyId, position, STATIC_CLIFF_CATEGORY, tileDef.colliderSize);
 	}
